feat: Add sound logic
