+++ PERONOME/AppDelegate.swift
@@ -16,8 +16,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
         
-        
-        
         window!.makeKeyAndVisible()
         window!.rootViewController = MetronomeViewController()
         return true
